# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Documentation for the Bookstore API",
        "contact": {
            "email": "admin@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/pembayaran": {
            "get": {
                "tags": [
                    "pembayaran"
                ],
                "summary": "Menampilkan semua pembayaran",
                "operationId": "b59db8649f14a4f1c15060ab304a629b",
                "responses": {
                    "200": {
                        "description": "List of Pembayaran",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "pembayaran"
                ],
                "summary": "Membuat pembayaran baru",
                "operationId": "edddbd797bca437acb54ce1f2b046691",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_pesanan",
                                    "id_pelanggan",
                                    "metode_pembayaran",
                                    "status_pembayaran",
                                    "tanggal_pembayaran",
                                    "total_pembayaran"
                                ],
                                "properties": {
                                    "id_pesanan": {
                                        "type": "integer"
                                    },
                                    "id_pelanggan": {
                                        "type": "integer"
                                    },
                                    "metode_pembayaran": {
                                        "type": "string"
                                    },
                                    "status_pembayaran": {
                                        "type": "string"
                                    },
                                    "tanggal_pembayaran": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "total_pembayaran": {
                                        "type": "number",
                                        "format": "float"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pembayaran created successfully"
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/pembayaran/{id}": {
            "get": {
                "tags": [
                    "pembayaran"
                ],
                "summary": "Menampilkan detail pembayaran berdasarkan ID",
                "operationId": "d32380c171eaf74c989e66f61e2e1aa5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pembayaran found"
                    },
                    "404": {
                        "description": "Pembayaran not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "pembayaran"
                ],
                "summary": "Mengupdate data pembayaran",
                "operationId": "4debc47d41004fa1272c6f9e244fc1fd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id_pesanan": {
                                        "type": "integer"
                                    },
                                    "id_pelanggan": {
                                        "type": "integer"
                                    },
                                    "metode_pembayaran": {
                                        "type": "string"
                                    },
                                    "status_pembayaran": {
                                        "type": "string"
                                    },
                                    "tanggal_pembayaran": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "total_pembayaran": {
                                        "type": "number",
                                        "format": "float"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pembayaran updated successfully"
                    },
                    "422": {
                        "description": "Validation errors"
                    },
                    "404": {
                        "description": "Pembayaran not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "pembayaran"
                ],
                "summary": "Menghapus pembayaran",
                "operationId": "70d6765ba080cb87bb03129e195fa424",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pembayaran deleted successfully"
                    },
                    "404": {
                        "description": "Pembayaran not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "bearerAuth": {
                "type": "http",
                "description": "Bearer token authentication using Laravel Sanctum",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
}